project migrated to maven;add ssdb pool
