+++ src/main/java/.gitignore
@@ -1,0 +1,8 @@
+*.o
+*.a
+*.class
+.DS_Store
+*.pyc
+*.swp
+*_cpy_*
+
