+++ src/main/java/README.md
@@ -1,0 +1,4 @@
+javassdb
+========
+
+SSDB Java client
