+++ src/main/java/com/udpwork/ssdb/SSDBException.java
@@ -1,0 +1,11 @@
+package com.udpwork.ssdb;
+
+public class SSDBException extends RuntimeException{
+	private static final long serialVersionUID = 849276316160517039L;
+
+	public SSDBException(String msg,Throwable e){
+		super(msg,e);
+	}
+
+	public SSDBException(){}
+}
